Fix MQTT_GetSubAckStatusCodes API for parameter type based on CBMC proof finding

+++ libraries/standard/mqtt/cbmc/proofs/MQTT_GetSubAckStatusCodes/MQTT_GetSubAckStatusCodes_harness.c
@@ -31,7 +31,7 @@ void harness()
 {
     MQTTPacketInfo_t * pSubackPacket;
     uint8_t * pPayloadStart;
-    uint16_t payloadSize;
+    size_t payloadSize;
 
     pSubackPacket = allocateMqttPacketInfo( NULL );
     __CPROVER_assume( isValidMqttPacketInfo( pSubackPacket ) );
